backend/usimc_2017/requirements.txt
cffi==1.9.1
cryptography==1.7.2
Django==1.10.2
django-crispy-forms==1.6.0
django-filter==1.0.1
djangorestframework==3.5.3
enum34==1.1.6
gunicorn==19.6.0
idna==2.2
ipaddress==1.0.18
jwt==0.5.2
Markdown==2.6.8
psycopg2==2.6.2
pyasn1==0.2.2
pycparser==2.17
six==1.10.0
typing==3.5.3.0
